+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Removed
+
+- Remove Opsgenie plugin.
+
 ## [0.37.2] - 2024-09-26
 
 ### Fixes